audio/alsa, audio/oss - Take the log level into account before logging a message.
The MSG macro in the ALSA and OSS OUTPUT code logs messages to the log file
of the module without taking the log level setting into account. Adjust the
MSG macro to log the message if the log level for the message is equal to, or
less than the log level setting.