audio/alsa, audio/oss - Assign all log messages a log level
All log messages that are logged using the MSG macro now have a log level
associated with them. The log levels that are currently assigned may not be
appropriate, however that can be tweaked at a later date. The documentation
also needs updating to indicate that the log level extends to audio output
activity logging.