server - More tightly integrate the server into a user session
This commit allows for tighter integration of the speech server into a user's
session. When the user logs in, the SPEECHD_PORT environment variable is set
to a unique port number, which is the base port set in intl/def.h, plus the
user's UID. When the server gets loaded, it then checks for the presence of
the SPEECHD_PORT environment variable, and uses that port for communication
if found. The environment variable will overwrite any setting in the config
file, but will not override the port command-line flag.

This integration is disabled by default, but can be enabled at compile time
with the --enable-session-integration command-line flag given to the
configure script. You can also set in which directory the shell fragment
gets placed, to allow it to be run at user login. This is set using the
--with-bash-profile-dir command-line flag. The default is
$sysconfdir/profile.d.