session - Add auto-spawn functionality as part of session integration
Add the ability for the C and python APIs to automatically load the
speech-dispatcher server, if it is not running. The spawn script  will determine
whether it is running as a user, or root, and will run speech-dispatcher
accordingly. If run as root, it will run speech-dispatcher using the system init
script for speech-dispatcher, specified at package build time, with the
--with-system-speechd-start-cmd argument.