Various fixes for deescape_dot.
First, this function used two different memory allocators: libc's malloc
and glib's g_malloc.
There's no way for the caller to know which allocator it was actually
using, since the decision depended on the contents of the argument string...
After this patch, deescape_dot always uses libc's malloc, since I'm
fairly certain that its return value is eventually handled by free
in all cases.

This function just needs to replace .. with . at the beginning of each
line.  See the definition of the speak command in the SSIP spec.  The previous
version was not quite doing that.

I also was able to remove a string duplication from the function which
calls deescape_dot, because the new implementation does not modify
its argument string.