memory leak
The _ibmtts_play function present in the file ibmtts.c has a loop to
remove and process commands present in playback_queue variable.
after the command is processed, the memory used by the command is released.
One condition to stop the loop is the content of the variable
ibmtts_stop_play_requested to be false.
When this happens a command could be removed but not processed, and the
memory used by the command is not released.
I tried to refactor the function slightly to avoid this situation

Also the module_init function allocates memory to the string info
but would not release at the end.