Do proper dot-stuffing in the Python bindings.
We need to escape the period, every time it occurs at the start of a line.
The old code had two problems.  First, it would not escape the second
period in the following sequence:
foo\r\n.\r\n.\r\nbar
Basically, it used the following replacement:
data.replace('\r\n.\r\n', '\r\n..\r\n')
This doesn't work when the replacement overlaps with the next string
to be replaced.  E.G., you have two consecutive lines consisting only of
period.  The second problem is that it is impossible to unambiguously
represent a line consisting of two dots.  The old code sends \r\n..\r\n
for both of the following inputs:
\r\n.\r\n
and
\r\n..\r\n

So we now follow the simple rule: always escape dot at the start of a line.