memory leaks
The _ibmtts_synth function calls  the ibmtts_next_part function to  get
the text to be synthesized, and a mark name, if any. After the text
is sent to the synthesizer, the memory occupied by it must be released.
However, if there was a mark name present, the memory occupied by the
text was not being released.
I tried to refactor the function so that the release of memory is
performed in a Single place.

In the function  ibmtts_search_for_sound_icon, the memory allocated to the
string filename, was not being released correctly if the file to be
searched is not present.