Add spd_audio_get_playcmd function.
All but a tiny part of this patch is from Andrei Kholodnyi.
It moves the determination of the play command into the plugins
themselves, rather than putting that information in
src/modules/generic.c.

Chris Brannon made the following additional changes in
src/modules/generic.c.
* Call module_audio_init_spd in module_audio_init, so that we
know which plugin we are using.
* Call spd_audio_close when necessary, as is done in other modules.
* Default to the "play" utility, when the proper play command can't
be determined.  play is widely available.
* Eliminate unnecessary string duplication when setting play_command.
play_command is now of type const char *.
It is either "play", or a constant string returned by
spd_audio_get_playcmd.
* The second and third arguments of the internal string_replace
function are now const char *, since they are never modified.

Signed-off-by: Christopher Brannon <cmbrannon79@gmail.com>
Reviewed-by: Hynek Hanke <hanke@brailcom.org>