Make autospawn optional for clients.
If not specified, it defaults to system default. We can't however use system default
from compile time to decide whether to use it, as in some cases, the library
can also be used to connect to different servers and it would be a mistake to autospawn
a local Speech Dispatcher while we are actually connecting to a remote one.