Autospawn mechanism fine-tuning
Due to reasons bellow, the autospawn script was eliminated and the
correct way now is to use the speech-dispatcher process itself with
the --spawn parameter. In this case, Speech Dispatcher will look into
its configuration, and if it cannot find a ''DisableAutoSpawn'' parameter,
then it will start, otherwise it will exit with an error value of 1.

The ~/.speech-dispatcher/ directory and configuration files are no longer
required to exist before speechd is started. If they don't exist, Speech
Dispatcher will use the system configuration from /etc/speech-dispatcher/
and will create this directory including the necessary log and pid subdirs.

Speech Dispatcher running from /etc/init.d/ is no longer recommended
and there is no longer any direct support for it. The current setup
is however general enough, so that you can setup an /etc/init.d/speech-dispatcher
script if you wish, by specifying the pid, log and conf paths on command line
on startup.

There is no way from the client to determine whether the user wants to start the ,,user
session'' or the ,,system'' Speech Dispatcher. So far, this has been determined by uid==0, but
this is not a good distinction. Speech Dispatcher should generally never run under root.  On
Debian systems, it is, as a system service, so far started under the special user
`speech-dispatcher', whose uid is not 0. For that reason, the mechanism was modifyed so that
autospawn always starts Speech Dispatcher under the current user.

Another reason for the elimination of the special autospawn shell script is that
it needed to perform a duplicate determination of the correct speechd configuration
file to check for DisableAutoSpawn. When this is done inside Speech Dispatcher itself,
the same logic is used.