flite_set_voice should be a no-op.
The previous version of flite_set_voice had the following behavior.
If the male1 voice was requested, it freed the voice data and
re-registered the same voice that was registered in module_init.
Otherwise, it did nothing.

This was not correct.
We only support one voice, and that voice is established in module_init.
The voice structure is managed by the flite library, and we shouldn't
try to free it.

Signed-off-by: Hynek Hanke <hanke@brailcom.org>