Change the way that lines are read in the server.
For the purpose of SSIP, the line terminator is carriage return linefeed.
The code that reads lines of SSIP should not accept a bare linefeed as
the line terminator.
If it does, then it will accept foo\n.\r\n as a complete message, even
though the end-of-data marker is \r\n.\r\n

Thanks to Trevor Saunders for suggesting this fix.

Signed-off-by: Hynek Hanke <hanke@brailcom.org>