Default values for configuration fixed.
Formerly, whether to provide a default value was being decided based
on the configuration variable having still a NULL value. Now this is
decided uniformly for integers and strings based on the value of the
corresponding variable_set parameter, which is more general in case
we want to influence it as in the case of the --spawn mechanism.