If --spawn is set, check connection parameters.
The port, socket_name and communication_method parameters
do not influence server startup if in autospawn mode,
but instead are checked to determine if the spawn startup
based on the configuration files will create an instance
of the server which will satisfy these parameters. Otherwise,
log spawn communication mismatch and exit with error code 1.