Autospawn checks whether spawn attempt will establish a connection.
Autospawn will pass expected connection parameters to Speech
Dispatcher on autospawn and Speech Dispatcher will issue an
error if this is not compatible with it's settings. Thus
we are sure that we are not autospawning misconfigured
Speech Dispatchers just due to user mistakes in client
configuration.

If Speech Dispatcher can't autospawn, a SpawnError is raised
containing the reason as reported by Speech Dispatcher on stderr.

Unified names for communication_method, socket path.
ConnectionMethod deprecated and renamed to CommunicationMethod
as in Speech Dispatcher.