Improved connect/spawn error reporting and proper inet spawn.
If attempt to connect to Speech Dispatcher fails, the
SSIPCommunicationError raised now contains a detailed description both
on the former error and on errors in attempts to resolve the problem
(such as autospawning the server). This will also include information
reported from the server if it can't start for a reason of it's own.

Additionally, autospawn is now attempted only for connections that are
determined to lead on localhost. If the connection is identified as
remote, autospawn of a local server is not attempted.