Fix growing mem requirements due to storing connection information
Previously, the settings information for all past connection
was stored and never released. The reason was pending messages
which are not yet spoken at the time of client disconnect and
history.


Since history is now disabled, we will completely delete every
connection as soon as the last message related to it is either
spoken or discarded.

This is a hack however. If history is to be re-implemented
in the future, connection information should be stored in
a pool of limited size and only removed (together with it's
messages) if the pool is full and space is needed for newer
connections.

Moreover, with the growth of the TFDSet structure, the current
implementation becomes difficult to maintain and should be
refactored.

This patch also fixes several other bugs on the propagation
of audio output settings.