Comment out less frequently used modules.
Active modules currently produce error messages
in the logfil if they loading fails, mostly due
to the particular synthesizer not being installed
on the system.

This can create confusion when looking for real
error messages in the logfiles, so it's better
to only allow the most widely used engines and let
people edit the file if they want to use another one.

For the future, the system of loadup of the
modules should be reworked.