Fix a segmentation fault when used with libao 1.0.0.
The libao audio library uses a structure named ao_sample_format.
Version 1.0.0 added a new element to that structure, named "matrix".
matrix should be NULL if the programmer doesn't care about it.  In
our case, it is unimportant, so NULL is the right value.
We now initialize stack-allocated variables of type ao_sample_format to
all-zeros, so that fields such as matrix are guaranteed to be zero.
Prior to this commit, matrix was uninitialized, thus causing libao to
crash.