Prevent deletion of wrong client from fd_settings during speak().
The code is supposed to delete the client associated with
current_message when it has no outstanding messages.
However, several statements referred to the "message" variable when
they should refer to current_message instead.
Also, the function client_has_messages() does not account for the
message that was just retrieved from the queue, so we also need to
check that current_message->settings.uid != message->settings.uid.

Signed-off-by: Hynek Hanke <hanke@brailcom.org>