Use hash table instead of SpeechdSocket array in server.
Formerly, the SpeechdSocket global variable was an array
to store information on open sockets to clients, in the
form of TSpeechDSock structures. The array was growing
with the number of connections in the past and contained
empty fields for socket numbers which were not used
or used in another way.

The global store for the TSpeechDSock informations
is now a hash table with keys by the file descriptor.
It's to be initialized, accessed and freed only using
wrapper functions.