Increase the range of eSpeak's volume.
To my ears, the espeak module was too quiet, even at the maximum volume.
Formerly, espeak_set_volume set the synthesizer's volume to a value in
the range 0 to 100.  Now, it uses a range of 0 to 200.
These values are consistent with those given in the help message for
the espeak command-line program, even though they contradict the ones
given in the library's header file.

Reviewed-by: Hynek Hanke <hanke@brailcom.org>

The eSpeak maintainer confirmed that 100 is the default
and <0,200> is a correct range.