Disable audio fallback and use Pulse Audio by default
Audio fallback is the right solution, but a prerequisite
is proper freeing of resources when the given instance
of server is not in an active session. Since this
is currently not implemented, audio fallback leads
to problems.

Under the previous settings "pulse, alsa", if
Speech Dispatcher is started in a session where
Pulse is not available, it will open ALSA, thus
preventing Pulse to start in all other sessions.
This subsequently breaks all audio for all
users although the session which blocks the
device is actually not active.

To prevent this, audio fallback is not used
by default.

Additionally, mention libao, which was missing.

Reviewed-by: Christopher Brannon <cmbrannon79@gmail.com>