remove double include from speechd.conf
There was an extra include directive in speechd.conf, which was also the
reason this file was being generated from speechd.conf.in.  If this
directive was in place and the file was installed as the system
configuration and not a user level configuration, the client settings
were being included twice.
If the system level configuration includes the files in clients/, that
will cover the default settings.  If a user wants to include the system
level client settings, they should add an include directive to their
user level configuration file to do so.  By default, spd-conf copies the
client settings over to the user's configuration, so they will not need
to do this.

Reviewed-by: Christopher Brannon <cmbrannon79@gmail.com>