Reopen libao device handles when audio parameters change.
Audio parameters can change from one track to another.  E.G., a
fragment of speech might not have the same sample rate as a sound icon.
We need to reopen libao's ao_device * handle when such changes occur.

Reviewed-by: Hynek Hanke <hanke@brailcom.org>

Additional note: Rate changes not only because of sound icons, but
e.g.  when switching to a different voice in Festival. So this patch
is fully justified.