Added new authors and thanks into AUTHORS
New persons were added. In the previous version, the people listed as
AUTHORS were listed with brief listing of their accomplishemnts. This
is however somehow ambigous and hard to maintain, so these lines have
been removed and the section is now a plain listing of the authors
names and contacts.

Contacts were removed from the THANKS section.