create a library of common routines
Move the code in the intl/ subdirectory to src/common and convert it
into a convenience library so that it can be linked with the modules and
server.

Move the include files from intl/ to include/ so that they can be easily
included by the server and modules.

Rename intl/fdset.h to include/speechd_types.h since this better
describes the contents of the file.  These types are to be shared
between the server and modules.

Reviewed-by: Christopher Brannon <cmbrannon79@gmail.com>