Rework the do_list_voices function, and fix a memory leak.
Thanks to Jose Vilmar Estacio de Souza for sending the original patch
to fix the memory leak in this function.
We were using a GString to accumulate data, and we weren't
freeing the unused portion of that string.

I also eliminated needless copying of dynamic strings, and I
re-structured a for-loop to improve clarity.
We now ignore voices with a NULL in the name field, though this should
never happen.  Also, I fixed a typo: sens -> send.

Reviewed-by: Andrei Kholodnyi <Andrei.Kholodnyi@gmail.com>

Andrei added a check of the non-empty voice list