Enable silent autotools build.
This makes the buildlog from make much cleaner (kernel-like). To use,
execute `make V=0' instead of `make'. And if you like it so much that you
want to make it default (i.e. even when using plain `make'), pass the
--enable-silent-rules argument to configure.

This is a feature of automake 1.11 and is enabled only when present
(i.e. falls back gracefully on older automake).

You have to regenerate configure before first use using build.sh

Reviewed-by: William Hubbs <w.d.hubbs@gmail.com>