major update to configure script
This update is based on running autoscan against our source tree,
removing some obsolete autotools macros and updating the calls to
others to use modern syntax.

Also this update re-arranges the script in a more logical arrangement,
updates the way we check for optional components., and reworks the way
the audio subsystems are handled.

The libraries and cflags for the audio systems are now handled in
separate variables instead of being accumulated in SPD_AUDIO_LIBS.

Reviewed-by: Christopher Brannon <cmbrannon79@gmail.com>