use GList instead of GHashTable for the output_modules
there is no need to implement a hash table for the output_modules with
module names as keys.
output_modules keeps OutputModule variables where module name
defines as a parameter in it. We can at any time retrieve it.
Also we want to keep modules ordering in the same way it exists in conf file
and it is impossible to do with GHashTable.

Reviewed-by: Christopher Brannon <chris@the-brannons.com>