pico.c: fixed threading issues.
There was a small threading issue.  The "play" thread
could be sending audio to the audio device while we are trying to close
it.  Now, we close the audio device after joining the "play" thread.

For good measure, all pointer variables are set to NULL,
once we are done with them.

Modified-by: Andrei Kholodnyi <Andrei.Kholodnyi@gmail.com>
remove exit call from module_close, we'll do it in main

Reviewed-by: Andrei Kholodnyi <Andrei.Kholodnyi@gmail.com>