rework default audio output method
Change the command line argument to --with-default-audio-method, which
is a more clear name for it than --with-audio-output-method. since this
argument configures the default method.

Rework the default selection method so that the cut command doesn't have
to be used by creating a variable which will be set to the name of the
first audio method that is found then overridden if there is a
--with-default-audio-method argument on the command line.

Reviewed-by: Andrei Kholodnyi <Andrei.Kholodnyi@gmail.com>