Guarantee that an output module is loaded at most once.
Since we now support the autodetection of output modules,
it is feasible that a module could be both requested by the
configuration file and autodetected at the same time.
We now insure that the same module is not loaded twice.  See the
comments in the code for my well-justified definition of "same".

Reviewed-by: William Hubbs <w.d.hubbs@gmail.com>