Correctly use execlp to specify module image name.
The first argument to the execlp function is the filename of the file
containing the executable image that will be loaded by the dynamic
loader. The second argument to the function then seems redundant, since
it should be the same as the first argument (since it is what gets
passed as argv[0] to the resulting program). But it can be different
from the first argument, e.g. when launching a shell script: then the
first argument will be e.g. /bin/sh, since that is the executable image
file from the view of the operating system, but the second argument
(i.e. argv[0]) will be the name of the shell script (e.g. script.sh).
Therefore it makes sense to specify the same for the first and second
argument when no interpreter is in place.

This fixes display using ps aux - the module processes launched by
speech-dispatcher main process had empty COMMAND field, making them
hardly identifiable for user inspecting running processes.

Reviewed-by: William Hubbs <w.d.hubbs@gmail.com>