Allow the speech-dispatcher module bin dir to be configurable
The speech-dispatcher synthesizer module model is such that any executable
binary or script can perform speech synthesis duties for speech-dispatcher,
as long as it can communicate via the appropriate protocol and file
descriptors. This also extends to using i386 binaries on an amd64 system.

Many distributions now have a model in place that allows the easy use of i386
binaries on an amd64 install. By default, the speech-dispatcher module
directory is placed in ${libdir}, which defaults to a 32-bit or 64-bit library
directory on some distros. Given the above, it should be possible to provide
an architecture independant location for modules to be installed to, and used.