Activate ibmtts user dictionnaries
Activates support of user dictionaries in IBM TTS.

This patch allows the user to customize the pronunciation of words, beginning of words, abbreviations,...
It activates the multi-language user dictionary feature supported by IBM TTS.

For more details on how adding dictionary files, refer please to the paragraph "User dictionaries" in ibmtts.conf below.

Implementation:

The dictionary files are loaded when a new language is selected and when the text previously added to the ECI engine has been processed.
Two stages:

- When a language is set, the main thread stores this information in the new integer named locale_index_atomic.
  This integer stores the index of the selected language in the eciLocales array.

- When the synthesis thread receives any message, it calls the new function named ibmtts_load_user_dictionary.
  This function checks locale_index_atomic and if modified, loads the new set of dictionary files.

Reviewed-by: Luke Yelavich <themuso@ubuntu.com>