config - Add pico-generic.conf
Due to licensing restrictions, some distributions do not build the pico module
in their speech-dispatcher package. The pico-genericconfig will allow users
of these distros to use pico, with the caviot that the functionality and
performance compared to the official pico driver is reduced.

Thanks to Samuel Thibault <sthibault@debian.org> for the patch.