API - Clean up voice/voice_type inconsistancies
The C API has set/get voice_type, the SSIP API has set/get voice. Change the
SSIP API to be consistant with the C API.

The SET VOICE SSIP command is now deprecated, and will be removed in 0.9.
The server code does provide for a GET VOICE command, but it was not
documented, so it has been renamed as per above.