API - Clean up voice/voice_type inconsistancies
The C API has set/get voice_type, the SSIP API has set/get voice. Change the
SSIP API to be consistant with the C API. Also remove the SET VOICE SSIP
command.

The server code does provide for a GET VOICE command, but it was not
documented, so it has been renamed as per above.