modules/espeak-ng: Replace white space with underscores in voice names
Recent versions of espeak-ng use spaces in their voice names, which breaks
the over the wire protocol between modules and the server, and from server
to clients. Fixing this properly requires a protocol change, which will
be in 0.9. Replacing with underscores is a work-around since breaking
protocol in a stable release series seems unwise.