server: Optimize list building
GLib lists are nodes alone, referenced by the a pointer to the first
node.  This means that to append to a list, it has to walk the whole
set of nodes to find the last one.  This is heavily inefficient when
adding several items, so instead use the common idiom of adding at the
start and then reverse the list.

Reviewed-by: Luke Yelavich <themuso@themuso.com>