Use autoconf check to decide whether daemon() needs to be defined
Previously this checked for the __SUNPRO_C compiler define, which
was a poor choice, as it depends on the libraries linked with, not
the compiler used to build.

While Solaris 10 didn't have daemon(), Solaris 11 libc does define it,
so let autoconf figure out whether or not to build a local version.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>