espeak-ng-mbrola-generic: propose alternative command line
which just lets espeak-ng output audio.

Its advantage is to be audio output agnostic, which helps e.g. when using
libao as libao.c does not set PLAY_COMMAND (as it can speak to other audio
apps).

This is not the default, though, as this prevents to set the sound
volume from the client. Which is not that useful IMHO, but I can be
wrong.