Close the audio output after closing the module
This allows modules to still use the audio output in module_close().

This is especially important if module_stop() doesn't actually stop
the playback but simply posts a stop request (which the Pulse output
make awfully easy), as it allows the module to rely on the output
honoring the request for normal operation, and only perform careful
cleanup (including shutting down any audio remains) in module_close().

This actually fixes an intermittent crash in at least the Baratinoo
module, but other modules are likely affected.

Closes #169.