[Port ibmtts_dynamic_loading to the speech-dispatcher 0.8 branch]
ibmtts: load dynamically libvoxin.so (libibmeci.so for compatibility)

Helps to remove an explicit dependency in sd_ibmtts (which could be a
showstopper using e.g. the rpm package manager if the corresponding
package of the external lib is not installed).

The dlopen/dlsym code inspired from emacspeak
(linux-outloud/atcleci.cpp).