audio: add feed_overlap method
This allows synths to get back control a bit earlier, to be able to
report marks and still have the time to feed further audio.

Ideally, we'd feed in advance the audio that is after the mark, and report
the mark exactly when the audio backend finished playing the part before the
mark, but this would be way more involved, while 20ms anticipation is not a
practical concern.