symbols: Fix SSML support
SSML is not only composed of "'<>&, but also / and whatever shows up in mark
names, so it's not actually reasonable to make the regexps use entities.

Instead we can mask out SSML tags by using U+E0XY and replace entities
with normal text, so that the unmodified regexps can work.