kali: Fix build with newer Kali versions
Newer libKali versions lack linking on some of its dependencies, so we
have to explicitly request them to fix configure-time detection.  They
are already listed manually in the Makefile, so this part is OK.

We also have to define LINUX for kali.h to have all the definitions we
need.