pulse: Fix default latency to 10ms
In src/server/configuration.c the default value was still set to 100 bytes,
i.e. just a couple ms, which triggers a bug in pulseaudio
https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/issues/871

10ms latency is considered by HCI (Human-computer Interaction) research
as real-time, so defaulting to 10ms should be fine enough.

Also, it is better to specify this as ms rather than bytes since the
number of bytes would depend on the rate of the synth.

Fixes #198