clibrary3: wait for speech termination
When testing each and every voice variant, one may want to use a dummy
audio output that terminates very fast, so the test does not last for
hours. We should then avoid waiting for a whole second.

It is still useful to wait a second in the clibrary2 case, to exercice
canceling previous speech.