Make all modules provide a list of voices
webspeech and such APIs are very surprised if the list of available
voices is empty.

So even for modules which have only one voice (cicero, dummy), we should
return information about it (it's also useful to provide the language).

Also, if a module has really no voice and cannot work, it should just abort
on startup (espeak-mbrola, festival, generic, ibmtts, kali).

Let's also make the server try to get the list of voices, to rule out
modules which do not provide one, and thus avoid loading them and rather
fallback to modules which do not have the issue.

Fixes #759