pulse: Restore using the full asynchronous pulseaudio API
Using the simple API is indeed simple, but does not provide flexible control
over audio writing, and can notably lead to underruns. This is
essentially the same as before, except that we keep writing audio data
as much as possible, since we can now control interrupting it from
spd_pa_simple_flush.

What we want is:

- Reduce server-side pre-buffering to min_audio_length
  - So playback can start as soon as reasonably possible without risking
    underruns
- Push as much data to the server as possible
  - So we avoid underruns
- Be able to stop as quickly as possible

Inlining the pulse simple implementation is not that complex and provides
best flexibility to achieve that.

That will also allow use to use pa_stream_get_time etc. to get the
playback position for proper indexing.

Fixes #885