Finish rationalizing input synchronization between main thread and output thread
Better make output_read_reply set waiting_for_reply itself, it makes a
lot of sense and simplifies the code.

Better unlock the mutex while reading the message, so that in case we
got a reply or an event before the other party tries to get it, it will
find it without waiting.