Ignore client SSML mode for anything but TEXT messages
SSML doesn't make sense for anything else, and it can break processing
of other messages in some cases.

For example, if a symbol expansion leads to converting it to a TEXT
command, it then has to be valid SSML (including surrounding <speak>
tags), otherwise a module might not accept it.  If the client is not in
SSML mode it worked fine as the input was translated to SSML, but if
the client was in that mode the input was still assumed to already be
SSML, although it actually wasn't as it's not possible (nor makes
sense) to send SSML for e.g. CHAR or KEY.

This e.g. broke the Baratinoo module for such symbol expansions (e.g.
when "y" is getting translated to "igrèque" to workaround pronunciation
issues in some synthesizers).