make set_volume actually set the volume of the stream and fix a silly compiler error
the volume we get from speech dispatcher is between -100 and 100 according to other modules, so we have to clamp that between 0 and 1. The way we do this is by scaling that value to be between 0 and 200, then we divide by 200 to get a number between 0 and 1. This is outlined more generically in the function, with specific min_value and max_value variables, but that's the gist of it
The compiler error stemmed from the fact that in the previous commit, a field inside a struct has been set to null instead of being set at the level of the function parameter