move calculating of the stride to runtime where all the information is available, move the obtained value in state to be used later
Stride is what's refered to when talking about the amount of bytes in a frame of audio data. This was calculated in on_process, because that's where it was needed.
The problem with that is that the calculation requires the size of a sample, for example i16 for the signed 16bit integer format, i8 for signed 8bit integer format, which wasn't stored anywhere in state. We previously assumed that the size of a sample is always i16, which could have caused unknown and weird behaviour when the format was signed 8bit integer instead. At the time, I thought that's rare, but since this change can be made, and also since it was asked for in code review, here we are
There were two options to solve this, the first was to store the sample format in state, while the second is to store the stride in there, calculated in pipewire_open. The second option was chosen, because the stride seemns like useful information to have, for potentially more than just the on_process callback