make backend abort spd upon unfavourable conditions, such as out of audio buffers or out of pipewire shared memory blocks
This can happen in a lot of cases, for example:

* pw crashes
* we ran out of buffers, tryed to dequeue too fast in on_process, or if we simply run out of memory
* due to a bug somewhere, too many modules loaded perhaps, we exhaust our opened file descriptors limit, so we can't create the eventfd

it is much better to simply abort in this case, without trying to reconnect or whatever, because the spawn function of a lot of speechd using applications will restart the speech dispatcher server when that happens, at which point spd can try again. If this persists, the only thing the user can do is restart their computer, or undo problematic config changes they made to their audio server, which either way is outside our control.

Better we crash and burn upon unfavourable conditions like this, than return error conditions, which make spd be stuck with no speech, unable to restart either because in its view, nothing is wrong