make stream adapt to speech dispatcher's sample rate changes
This can happen, most regularly, if new output modules are loaded, and those send a different sample rate than what the stream is configured for using.
In such cases, at least in the current implementation, the stream gets disconnected entirely, to then be reconnected later in the same function with updated values.
Warning: if you use pw-link or jack based tools, or perhaps even obs, to link apps to one another, the link between the client where this happens and other apps will be lost, because changing of the sample rate triggers a stream distruction, which cancels all links

This can also happen because spd spuriously called this function when nothing changed, in which case we print a warning and return because there's nothing to be done.