allow the backend to be usable  with older versions of pipewire
The issue was that untill relatively recently, pipewire didn't indicate to applications how much data it expected from them, for the apps to then try and load that amount in a buffer they dequeue from on_process.
So, what this does is it completes the conditional compilation work done a while ago, by making it so that something useful happens even if the user runs a slightly older pipewire version.
What this does is similar to other applications back then, it pushes as much as it has in the ringbuffer the audio thread of spd fills up, regardless if in modern pipewire, that would xrun because it would be less than requested.

Furthermore, some edits were done, to disambiguate some variable names, and make people understand the meaning of that part of the code better, because it's complicated as it is, we don't need ambiguous variable names thrown in the mix.
To that extent, available_bytes was renamed to what it actually means, how many bytes are available from the spd side, and number_of_bytes was renamed to total_number_of_bytes_in_pipewire_buffer, to make sure people understand what's being calculated there