Do not add symbols from base/symbols.dic when they are not in <locale>/symbols.dic
E.g. ⇄ is in base/symbols.dic, but not in fr/symbols.dic, so we do not
want to take the definition from base/symbols.dic, and let emojis.dic
provide a definition.

This is the same goal as f9622994cbd43ce67f992f87337fbb0174792a21, but
without losing base/symbols.dic providing symbol default properties.